:recycle: move interface definitions to dedicated files
