+++ src/contracts/exportService/IExporter.ts
@@ -1,0 +1,9 @@
+export interface IExporter {
+  /**
+   * Exports the current diagram as the last defined format.
+   *
+   * @params filename Name of the file that will be exported.
+   * @returns the exported diagram.
+   */
+  export(filename: string): Promise<void>;
+}
