+++ src/contracts/exportService/index.ts
@@ -1,1 +1,4 @@
-export * from './IDiagramExportService';
+export {IDiagramExportService} from './IDiagramExportService';
+export {IExporter} from './IExporter';
+export {IExportSVG} from './IExportSVG';
+export {IExportXML} from './IExportXML';
